Ensuring an appender exists.
Prior to this commit, setting a level on a logger would implicitly set an appender. This
is not always the behaviour you want when setting a level. This commit removes the
implicit add of the appender, but ensures that there is at least one appender set in the
hierarchy when set-logger! returns. This means that setting levels will only set levels
in a complex configuration, but when there is nothing configured (the out-of-the-box case),
then the most desirable behaviour results: a single simple appender is added at the root.